Authentication.

+++ src/main/java/com/zuehlke/securesoftwaredevelopment/controller/PersonsController.java
@@ -2,13 +2,19 @@ package com.zuehlke.securesoftwaredevelopment.controller;
 
 import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
 import com.zuehlke.securesoftwaredevelopment.domain.Person;
+import com.zuehlke.securesoftwaredevelopment.domain.Role;
 import com.zuehlke.securesoftwaredevelopment.domain.User;
 import com.zuehlke.securesoftwaredevelopment.repository.PersonRepository;
+import com.zuehlke.securesoftwaredevelopment.repository.RoleRepository;
 import com.zuehlke.securesoftwaredevelopment.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -26,13 +32,16 @@ public class PersonsController {
 
     private final PersonRepository personRepository;
     private final UserRepository userRepository;
+    private final RoleRepository roleRepository;
 
-    public PersonsController(PersonRepository personRepository, UserRepository userRepository) {
+    public PersonsController(PersonRepository personRepository, UserRepository userRepository, RoleRepository roleRepository) {
         this.personRepository = personRepository;
         this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
     }
 
     @GetMapping("/persons/{id}")
+    @PreAuthorize("hasAuthority('VIEW_PERSON')")
     public String person(@PathVariable int id, Model model) {
         model.addAttribute("person", personRepository.get("" + id));
         model.addAttribute("username", userRepository.findUsername(id));
@@ -40,6 +49,7 @@ public class PersonsController {
     }
 
     @GetMapping("/myprofile")
+    @PreAuthorize("hasAuthority('VIEW_MY_PROFILE')")
     public String self(Model model, Authentication authentication) {
         User user = (User) authentication.getPrincipal();
         model.addAttribute("person", personRepository.get("" + user.getId()));
@@ -47,28 +57,57 @@ public class PersonsController {
         return "person";
     }
 
+    private boolean isAdmin(User user){
+        List<Role> roles = roleRepository.findByUserId(user.getId());
+        for (Role role : roles) {
+            if (role.getName().equals("ADMIN")) { return true; }
+        }
+        return false;
+    }
+
     @DeleteMapping("/persons/{id}")
-    public ResponseEntity<Void> person(@PathVariable int id) {
-        personRepository.delete(id);
-        userRepository.delete(id);
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
+    public ResponseEntity<Void> person(@PathVariable int id, Authentication authentication) {
+        User user = (User) authentication.getPrincipal();
+        if ( isAdmin(user) || user.getId() == id) {
+            personRepository.delete(id);
+            userRepository.delete(id);
+        } else{
+            throw new AccessDeniedException("Access denied.");
+        }
 
         return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/update-person")
-    public String updatePerson(Person person, String username) {
-        personRepository.update(person);
-        userRepository.updateUsername(Integer.parseInt(person.getId()), username);
-        return "redirect:/persons/" + person.getId();
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
+    public String updatePerson(Person person, String username, Authentication authentication) {
+        User user = (User) authentication.getPrincipal();
+        Person tmpPerson = personRepository.get("" + user.getId());
+        boolean isAdmin = isAdmin(user);
+        if (isAdmin|| tmpPerson.getId().equals(person.getId())) {
+            personRepository.update(person);
+            userRepository.updateUsername(Integer.parseInt(person.getId()), username);
+        }
+        else{
+            throw new AccessDeniedException("Access denied.");
+        }
+        if (isAdmin)
+            return "redirect:/persons/" + person.getId();
+        else
+            return "redirect:/myprofile";
     }
 
     @GetMapping("/persons")
-    public String persons(Model model) {
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
+    public String persons(Model model, Authentication authentication) {
         model.addAttribute("persons", personRepository.getAll());
+        model.addAttribute("currPerson", personRepository.get("" + ((User) authentication.getPrincipal()).getId()));
         return "persons";
     }
 
     @GetMapping(value = "/persons/search", produces = "application/json")
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     @ResponseBody
     public List<Person> searchPersons(@RequestParam String searchTerm) throws SQLException {
         return personRepository.search(searchTerm);
